Depende on importlib_metadata instead of `importlib.metadata`
This solves the following issue:

```
TypeError: entry_points() got an unexpected keyword argument 'group'
```

This comes due to an old implementation of `importlib` and specifically
when I'm running my app inside a virtual environment. Attempts to `pip3
install --force-reinstall importlib`, inside the virtualenv, did not
work since it seems that `importlib` is brought from the standard python
distribution regardless of whether the environment is sourced or not. To
accommodate running the app inside and outside a virtualenv I'm instead
depending on `importlib_metadata` which seems to resolve the
aforementioned issue.